removed various warning from Polyhedra compilation

+++ elina_poly/opt_pk_cherni.c
@@ -566,10 +566,10 @@ void opt_cherni_minimize(opt_pk_internal_t* opk,
   	}
   }*/
   //else{
-  for (i = 0; i < C->nbcolumns - 1; i++) {
-    F->p[i][i + 1] = 1;
-  }
-        /* conversion */
+  	for (i=0; (int)i<C->nbcolumns-1; i++){
+    		F->p[i][i+1] = 1;
+  	}
+  	/* conversion */
   	opk->exn = ELINA_EXC_NONE;
   	op->nbline = opt_cherni_conversion(opk, C, 0, F, satC, C->nbcolumns-1);
